Extract formatting functions from common.
