+++ snapcraft/internal/common.py
@@ -253,41 +253,3 @@ def get_pkg_config_paths(root, arch_triplet):
     ]
 
     return [p for p in paths if os.path.exists(p)]
-
-
-def combine_paths(paths, prepend, separator):
-    paths = ['{}{}'.format(prepend, p) for p in paths]
-    return separator.join(paths)
-
-
-def format_path_variable(envvar, paths, prepend, separator):
-    if not paths:
-        raise ValueError(
-            "Failed to format '${}': no paths supplied".format(envvar))
-
-    return '{envvar}="${envvar}{separator}{paths}"'.format(
-        envvar=envvar, separator=separator, paths=combine_paths(
-            paths, prepend, separator))
-
-
-def humanize_list(items, conjunction):
-    """Format a list into a human-readable string.
-
-    :param list items: List to humanize
-    :param str conjunction: The conjunction used to join the final element to
-                            the rest of the list (e.g. 'and').
-    """
-
-    if len(items) == 0:
-        return ''
-
-    quoted_items = ['{!r}'.format(item) for item in sorted(items)]
-    if len(items) == 1:
-        return quoted_items[0]
-
-    humanized = ', '.join(quoted_items[:-1])
-
-    if len(items) > 2:
-        humanized += ','
-
-    return '{} {} {}'.format(humanized, conjunction, quoted_items[-1])
